add: added UI for DNS settings
